Add support for GPIOs 11 (output) and 12 (input)
Add support for two GPIO pins:
Pin 11 is configured as an output.
Pin 12 is configured as an input with an active pullup.

A secondary timer is created to notify with the GPIO data.
Both this timer and the custom data timer are now 1 second.

Cleaned up some debug prints, minimal prints default for INFO level.